refactor(auth): fold headers.py into refresh.py (ADR-0033 sanctioned merge)
``_auth/headers.py`` was 68 lines holding exactly one function,
``_resolve_token_route_kwargs``, whose only three call sites are the
token-fetch entry points in ``_auth/refresh.py`` (plus the facade alias in
``auth.py``). It failed the deletion test as a standalone module: it existed
because the routing glue was too small to justify its own file and too
entry-point-specific to sit in ``_auth/account.py`` next to the authuser
helpers it composes — exactly the shallow-module shape ADR-0033 exists to
retire.

The function moves verbatim (including its function-local ``account`` /
``cookies`` imports) into a labelled section of ``refresh.py`` immediately
above its callers, and ``headers.py`` is deleted. No import cycle is created:
``refresh.py`` already imported ``_auth.account`` at module scope.

``notebooklm.auth._resolve_token_route_kwargs`` keeps resolving to the SAME
function object — the facade alias is re-pointed at ``_auth.refresh`` in the
same commit, so the identity the public-surface manifest pins is preserved.

Guardrail bookkeeping (same commit, per the plan's rule 2):
* the manifest's headers clause is REWRITTEN, not re-run: the identity
  assertion now targets ``refresh`` (plus a ``__module__`` check that the
  function is genuinely defined there and not re-aliased), and the two dead
  module-existence assertions are dropped rather than re-pointed — module
  existence is already covered by the seam-module clause;
* ``seam_modules`` shrinks by ``headers`` (shrink-only, never grows);
* ``docs/architecture.md``'s ``_auth`` index loses the ``headers.py`` row in
  all three places it appears (module table, flat index, repository tree) and
  ``refresh.py``'s row absorbs it.

The sanctioned-merge ceiling is deliberately NOT added here. ADR-0033 pins at
MEASURED LOC with zero headroom, and this PR grows ``refresh.py`` twice more
(the ``_cold_fallbacks`` extraction and the deps record), so the module crosses
the budget once and takes exactly one pin, measured at the PR tip. The
module-size ratchet is therefore red on this commit and green at the tip.
