+++ docs/architecture.md
@@ -528,7 +528,6 @@ the default dependency.
 | [`_auth/storage_writer.py`](../src/notebooklm/_auth/storage_writer.py) | **Shim** — re-exports the writer API from `_auth/storage.py`; defines nothing. Removed at the next major. |
 | [`_auth/storage_transaction.py`](../src/notebooklm/_auth/storage_transaction.py) | **Shim** — re-exports `in_storage_transaction` + the lock-failure policies from `_auth/storage.py`; defines nothing. Removed at the next major. |
 | [`_auth/extraction.py`](../src/notebooklm/_auth/extraction.py) | Cookie/token extraction from browser sessions. |
-| [`_auth/headers.py`](../src/notebooklm/_auth/headers.py) | HTTP header construction. |
 | [`_auth/cookies.py`](../src/notebooklm/_auth/cookies.py) | Cookie maps + `_update_cookie_input` helper. |
 | [`_auth/cookie_policy.py`](../src/notebooklm/_auth/cookie_policy.py) | Domain allowlist, cookie-domain builder (`build_cookie_domain_allowlist`), and cookie policy decisions. |
 | [`_auth/cookie_semantics.py`](../src/notebooklm/_auth/cookie_semantics.py) | Shared cookie-shape and expiry semantics used by sanitized auth loaders and persistence boundaries. |
@@ -544,7 +543,7 @@ the default dependency.
 | [`_auth/single_flight.py`](../src/notebooklm/_auth/single_flight.py) | Cross-loop single-flight coalescing core (c-PR2; [ADR-0030](./adr/0030-one-recovery-ladder.md)): a process-global `(canonical path, rung policy)` flight registry (leader `asyncio.Task` mirrored into a `concurrent.futures.Future`; followers on any loop bridge via `shield(wrap_future)` + settle-before-propagate) plus a per-canonical-PATH `success_epoch` (relocated `_REFRESH_GENERATIONS` semantics). Consumed by `refresh.py` (refresh-cmd) and `recovery.py` (cold/L4). One `threading.Lock` guards only brief claim/registry/epoch sections, never across an await. |
 | [`_auth/account.py`](../src/notebooklm/_auth/account.py) | Account profile + multi-account switching. |
 | [`_auth/session.py`](../src/notebooklm/_auth/session.py) | `refresh_auth_session(auth=..., kernel=..., auth_coord=..., lifecycle=..., cookie_persistence=...)` implementation called by `AuthRefreshCoordinator`. Takes five explicit keyword-only collaborators instead of a Session-shaped owner Protocol; the previous `RefreshAuthCore` Protocol and the `update_auth_tokens` / `update_auth_headers` Session-level forwards have been removed. |
-| [`_auth/refresh.py`](../src/notebooklm/_auth/refresh.py) | Token refresh driver (external login command, secret redaction). Coalesces refresh-cmd runs across loops via the `single_flight` core and serializes the subprocess across processes with a per-path refresh flock ([refresh-2]). |
+| [`_auth/refresh.py`](../src/notebooklm/_auth/refresh.py) | Token refresh driver (external login command, secret redaction). Coalesces refresh-cmd runs across loops via the `single_flight` core and serializes the subprocess across processes with a per-path refresh flock ([refresh-2]). Owns the cold-start fallback **sequence** in one place (`_cold_fallbacks`: L3/L4 cold recovery, then the L2.5 refresh-cmd rung) while `recovery.py` keeps the flight-leader body. Also owns `_resolve_token_route_kwargs`, the token-route glue absorbed from the deleted `_auth/headers` module (ADR-0033 sanctioned merge — it held one function whose only call sites are the entry points here). |
 | [`_auth/keepalive.py`](../src/notebooklm/_auth/keepalive.py) | Cookie keepalive + `__Secure-1PSIDTS` rotation. |
 | [`_auth/psidts_recovery.py`](../src/notebooklm/_auth/psidts_recovery.py) | Inline PSIDTS recovery for cold-start (see issue #865). Invoked from the loader **wrapper** bodies, never the network-free pure loader ([ADR-0030](./adr/0030-one-recovery-ladder.md)); its env-var check shares `paths.resolve_auth_json_env`. |
 | [`_auth/master_token.py`](../src/notebooklm/_auth/master_token.py) | Headless master-token auth: minting primitives (exchange/mint/persist) PLUS the whole audited transaction (`bootstrap_from_oauth_token`, `remint_from_stored_token` — the shared kernel L4's re-mint and the CLI's operator refresh both call, `bootstrap_storage_from_master_token` returning a `BootstrapOutcome` kept internal, `bootstrap_missing_storage_from_master_token` collapsing it to the bool the CLI actually crosses the boundary with, `assert_account_writable`), relocated here from the CLI by #2103's PR-2 structural follow-up (ADR-0023). |
@@ -999,7 +998,6 @@ Per-file index plus the full `src/notebooklm` + `tests` repository tree. The tre
 | `auth.py` | Authentication facade — **almost pure re-exports** (the only remaining function body is `async def enumerate_accounts`, which binds `_poke_session` as a default dependency; ADR-0003 records the optional-`async` audit command). Every other top-level name forwards from the relevant `_auth/*` module: `auth._validate_required_cookies` is identity-equal to `_auth.cookie_policy._validate_required_cookies`, and `load_auth_from_storage` / `AuthTokens` live in `_auth/tokens.py`. **ADR-0003's flat-re-export goal was closed by ADR-0014.** Tests that need to rebind policy names patch `_auth.cookie_policy.X` directly. |
 | `_auth/paths.py` | Storage paths and filesystem helpers |
 | `_auth/extraction.py` | Cookie/token extraction from browser sessions |
-| `_auth/headers.py` | HTTP header construction |
 | `_auth/cookies.py` | Cookie map manipulation + `_update_cookie_input` |
 | `_auth/cookie_policy.py` | Cookie-domain allowlist, `build_cookie_domain_allowlist` builder, and policy decisions |
 | `_auth/cookie_semantics.py` | Shared cookie-shape and expiry semantics at loader/persistence boundaries |
@@ -1183,7 +1181,6 @@ src/notebooklm/
 │   ├── __init__.py
 │   ├── paths.py                 # Storage paths and filesystem helpers
 │   ├── extraction.py            # Cookie/token extraction from browser sessions
-│   ├── headers.py               # HTTP header construction
 │   ├── cookies.py               # Cookie maps + _update_cookie_input
 │   ├── cookie_policy.py         # Domain allowlist + cookie-domain builder and policy
 │   ├── cookie_semantics.py      # Shared cookie-shape and expiry semantics
